Stats view

+++ Loopover/App/Constants.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct Constants {
     static let letters = Array("ABCDEFGHIJKLMNOPQRSTUVWXYZ")
-    static let chartKeyPaths: [KeyPath<Solve, Int>] = [\.numMoves, \.timeInHundreths]
+    static let chartKeyPaths: [KeyPath<Solve, Int64>] = [\.numMoves, \.timeInHundreths]
     static let sizes = [3, 4, 5, 6, 7]
 }
